Merged the revapi-api module with revapi + random fixes here and there.
The main class will no longer be in the org.revapi.Revapi,
because the standalone module is going to be more complex and would
add too much deps to the revapi module.

Fixed primitive array handling in the ClassTreeInitializer.

Beefed up the text reporter to be of some use actually.

api analyzers, problem tranforms, element filters and reporters are
now closeable so that they have a chance to clean up after analysis.