Integrate Checkstyle
This patch introduces an integration with the Checkstyle[1] tool for
static code analysis.

The introduced configuration is a very basic one which has been tested
and proven to pass the build with the given source code.
Following patches will suggest additional checks enabled and fix the
source code where it fails these checks.

[1] http://checkstyle.sourceforge.net/