Improve the transformation handling.
Before, when 1 transform wanted a diff removed but other transformed,
the transformation "won" over the removal. If more than one transformation
wanted the difference transformed, only the last transformation prevailed.

I think that is wrong.

Now when a difference is deemed to be removed by any transformation it is
removed. On the other hand if more than 1 transformation applies to a
difference, then all of the transformed differences are taken into account.