[#35] Move class and package filter right into API tree construction logic.
This is necessary so that we can be clever and correctly not drag the types
that are used by the ignored classes into the API. That would be the case
if the filtering was done at the filter stage (as it was done originally)
because the API tree would already be constructed and the "drags"
determined. It would then be much harder/more expensive to figure out if a
class is only dragged into API by an ignored type.