[#41] More robust handling of methods moved between types. 4 new reported problems were added: * method moved to superclass * inherited method moved to class * method now abstract * method no longer abstract
"overriding method removed" problem was removed and is now reported
as "method "method moved to superclass".

The former 2 new problems account for the cases where a method moves in
the type hierarchy, while the latter two were missed before and were added
to account for 1 case that would also best describe a situation when
an abstract method moved in a particular way between versions.