Introduce new way of scanning the classpath.
Instead of ASM-based crawling, use the StandardJavaFileManager and
javax.lang.model classes directly. This seems to be moderately faster
than what we had before.

This is functionally equivalent to the previous codebase but nevertheless
it is still failing to either find breakage in some cases or in other cases
it reports false positives.

This will need a complete revamp of the "member accessibility" problem.
Instead of trying to figure out if members of a superclass are somehow
accessible through subclasses, it will be easier to check all methods,
including the inherited ones, at each class. This is quite huge though,
because methods, etc. can no longer be represented by their Elements in
that case, but need to be represented by concrete TypeMirrors.

Fun times ahead...