Added a new "equivalent" difference severity.
This severity means that there is no perceptible change in behavior
because of the difference, which is different from non-breaking.

Allow equivalent changes in patch version increases in semver.

Update the severities of the java extension to use this new severity.

Update the docs to include all reported difference codes.

Clarified some differences.