[#34] - Change the way extensions are loaded and configured.
Now Revapi instantiates and the extensions on its own and manages their
lifecycle. Each analysis run instantiates all the extensions anew.

There is a new type of configuration. Instead of old JSON object where
each extension could declare multiple "places" it reads configuration
from, it is now ALSO possible to declare configuration as a JSON
array which contains objects with 2 keys: "extension" and
"configuration".

Each configurable now declares a single "extensionId" (corresponding
to the "extension" in the configuration JSON) and declares a single
JSON schema that can configure it (that the value of the "configuration"
configuration key must conform to).

Because the old-style configuration is an object and new-style
configuration is an array, we can keep support for both.