[#19] Make configuration merging interoperable with old and new-style config.
This is done by converting to new-style directly in AnalysisContext builder
(which is now the sole way of instantiating the AnalysisContext).

Also give up on backwards compatibility for extensions reading
the configuration. We now have a superior way of dealing with config which
supports both multi-instance and XML conversion, so it's time to switch ;)