#114 Inherited methods are assigned the archive of the target type, not the source type.
#109 changed the decision logic on when to descend into a filtered out
and/or element exclusive to just one of the APIs. As a consquence of
that a bug was uncovered in revapi-java which assigned the archive of
the original method definition to elements representing the methods in
classes that inherit from the defining class.

If such methods were defined in archives that were not excluded but
the inheriting class was filtered out by an archive filter, the methods
would still be API checked because the archive filter would think they
come from a included archive.