#112 - Fix finding types by binary name while not making it much slower.
While the common case for java classes is slightly slower (for the
nested classes), it is several orders of magnitude faster for the
"degenerate" cases like classes generated by Scala which contain many $
in the class names.

It also now can handle hopefully all corner cases which were not handled
by the previous implementation.