Do not skip transformed difference in cases of multiple transformers.
In the case when the configuration had multiple transformers it was possible
to skip transformed difference in case if any transformer
after actually transforming transformer returns null as the difference.
A described situation can happen, for example, in the case when the configuration has
SemverIgnoreTransformer and ClassificationTransfromer defined.

This PR also fixes a bug in classification-schema.json where `BINARY` value was incorrectly described.