#123 rework overload matching
The difference is now done using Levenshtein distance and the comparison
is stable and can handle the cases where more than 1 pairing is possible
with equal difference in types. Pure return-type changes are prefered over
parameter changes when declaring methods correspondence. I.e. we going to
rather report a return type change than a parameter type change if both
are possible to report.