Add the new revapi.versions extension and deprecate revapi.semver.ignore.
For that to make sense add a new method to Archive.Versioned to be able
to find out the version-less name of the archive.

Support the notion of references between elements directly in the main API.
This is used in the revapi.versions to figure out the "true" archive in the
primary API that makes the API changes public.