Fixes #222, #242.
The method parameters of inherited generic methods are correctly assigned
the model representation (type mirror) from the executable type viewed as
a member of the target type.

When looking at return types and parameter types, we also recognize the
differences in the erased signatures to correctly report possible binary
breakage due to different binary method signatures.